\name{nilsTCGA}
\alias{nilsTCGA}
\alias{nilsTCGA}
\docType{package}
\title{
\packageTitle{nilsTCGA}
}
\description{
\packageDescription{nilsTCGA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nilsTCGA}
\packageIndices{nilsTCGA}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{Nils Weng}

Maintainer: \packageMaintainer{Nils Weng}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
